+++ Delta.CapiNet/Delta.CapiNet/Capi32.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using System.Collections.Generic;
-using System.Runtime.InteropServices;
 using System.Security.Cryptography.X509Certificates;
 
 using Delta.CapiNet.Internals;