+++ Delta.CertXplorer/third-party/Delta.CapiNet/Certificate.cs
@@ -9,7 +9,7 @@ namespace Delta.CapiNet
     /// </summary>
     public class Certificate
     {
-        private X509Certificate2 innerObject = null;
+        private X509Certificate2 x509 = null;
 
         /// <summary>
         /// Initializes a new instance of the <see cref="Certificate"/> class.
@@ -17,15 +17,17 @@ namespace Delta.CapiNet
         /// <param name="certificate">The certificate.</param>
         public Certificate(X509Certificate2 certificate)
         {
-            innerObject = certificate;
+            x509 = certificate;
         }
+
+        #region Properties
 
         [Browsable(false)]
-        public X509Certificate2 X509Certificate2
+        public X509Certificate2 X509
         {
-            get { return innerObject; }
+            get { return x509; }
         }
-
+        
         /// <summary>
         /// Gets a value indicating whether this certificate is valid.
         /// </summary>
@@ -39,8 +41,44 @@ namespace Delta.CapiNet
             get
             {
                 var now = DateTime.Now;
-                return now >= innerObject.NotBefore && now <= innerObject.NotAfter;
+                return now >= x509.NotBefore && now <= x509.NotAfter;
             }
         }
+
+        #region X509Certificate2 forwarding properties
+
+        /// <summary>
+        /// Gets the subject distinguished name of this certificate.
+        /// </summary>
+        public X500DistinguishedName SubjectName
+        {
+            get { return x509.SubjectName; }
+        }
+
+        /// <summary>
+        /// Gets a value that indicates whether this certificate contains a private key.
+        /// </summary>
+        public bool HasPrivateKey
+        {
+            get { return x509.HasPrivateKey; }
+        }
+
+        public X500DistinguishedName IssuerName 
+        {
+            get { return x509.IssuerName; }
+        }
+
+        /// <summary>
+        /// Gets or sets the associated alias for a certificate.
+        /// </summary>
+        public string FriendlyName
+        {
+            get { return x509.FriendlyName; }
+            set { x509.FriendlyName = value; }
+        }
+
+        #endregion
+
+        #endregion
     }
 }