+++ Delta.CertXplorer/third-party/Delta.CapiNet/Properties/AssemblyInfo.cs
@@ -10,5 +10,10 @@ using System.Runtime.InteropServices;
 [assembly: AssemblyTrademark("")]
 [assembly: AssemblyCulture("")]
 [assembly: ComVisible(false)]
-[assembly: AssemblyVersion("1.1.0.0")]
-[assembly: AssemblyFileVersion("1.1.0.0")]
+[assembly: AssemblyVersion(AssemblyInfo.Version)]
+[assembly: AssemblyFileVersion(AssemblyInfo.Version)]
+
+internal static class AssemblyInfo
+{
+    public const string Version = "1.1.1.0";
+}